NCBC-1613: Config processing thread is never terminated
Motivation
----------
The config thread uses a BlockingCollection as a queue; the
BlockingCollection is never emptied so the thread could block there,
waiting for more items.

Modifications
-------------
Update the Dispose method so that the queue is flagged to not accept any
more items. Then remove any items that haven't been processed, finally
call Dispose on the BlockingCollection.

Result
------
The config thread should safely exit because the BlockingCollection is no
longer waiting for new items.

Change-Id: Ib4123afd25713c91ef7c7a6edeb537bcc69a29fb
Reviewed-on: http://review.couchbase.org/87351
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>