NCBC-1616: Add per-operation tracing (OpenTracing)
MOTIVATION
----------
It is desirable for operations of all types to record granular
information to help diagnose and solve performance issues. An open
source solution is OpenTracing, which describes an API and describes how
span structures can be recorded and organised.

This change introduces the OpenTracing API dependency and attempts to
solve a common scenario of Slow Operation tracking.

More information on OpenTracing can be found here: http://opentracing.io/

MODIFICATIONS
-------------
- Add the OpenTracing API nuget dependency
- Add the ThresholdLoggingTracer implementation with supporting
  configurations options
- Add basic implementations of other required OpenTracing strucutures,
  such as SpanBuilder, Span and SpanContext.
- Extend each operation type (KV, View, N1QL, FTS and Analytics) to
  maintain an active span property
- Extend each service layer to track encoding, dispatch and decoding of
  each operation
- Provide additional testing and fix broken unit tests

RESULT
------
An OpenTracing complication ThresholdLoggingTracer is used to track and log
slow operation at a given interval. The SDK has been upgraded to track
and record Spans for discrete phases of an operation life cycle.

There are a number of configurations available to control service
thresholds, sample size and log interval. The tracing feature can also
be disabled entirely with a NullTracer being used instead which does not
record any activity.

NOTE: The OpenTracing implementation is not committed and my change in
the future.

Change-Id: Ie57df90860f94cb813ca7a538693aac28aa9db02
Reviewed-on: http://review.couchbase.org/87536
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>