NCBC-1617: Make SslConnection check IsEncrypted and IsSigned after authentication
Motivation
----------
Although an exception is thrown if authentication fails for an SSL/TLS
connection, the MSDN documentation states:

"Before sending or receiving information, the client and server should
check the security services and levels provided by the SslStream to
determine whether the protocol, algorithms, and strengths selected meet
their requirements for integrity and confidentiality. If the current
settings are not sufficient, the stream should be closed. You can check
the security services provided by the SslStream using the IsEncrypted and
IsSigned properties. The following table shows the elements that report
the cryptographic settings used for authentication, encryption and data
signing."

https://msdn.microsoft.com/en-us/library/system.net.security.sslstream(v=vs.110).aspx

Modifications
-------------
Check the IsEncyrpted, IsSigned properties, as well as the
IsAuthenticated property after authenticating the SslStream and throw an
AuthenticationException if any are false.

Result
------
The SslStream will fail if IsEncrypted, IsSigned, or IsAuthenticated is
false.

Change-Id: I56332bc5a2a33daa74d6a6d7fb5d39a26d07f996
Reviewed-on: http://review.couchbase.org/87883
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>