NCBC-1616: Send connection ID to HTTP services in custom header
MOTIVATION
----------
When sending requests to the HTTP services (N1QL, FTS, etc), the
connection ID should also be sent in the headers for additional context.

NOTE: These headers are not yet used, but may be in the future.

MODIFICATIONS
-------------
- Add utility method to ClientIdentifer to format the Instance and
  connection IDs into the correct format
- Update HTTPServiceBase to generate random ulong instad of GUID per
  service
- Update IOServiceBase to reuse FormatConnectionId method instead of
  copying it
- Update tests to verify behaviour

RESULT
------
When HTTP servies send requets, a connection ID HTTP header is included.

Change-Id: Ifdf8bf9625e348debba3464052e11be09c419119
Reviewed-on: http://review.couchbase.org/88503
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>