NCBC-1628: Add support for N1Ql raw parameters
MOTIVATION
----------
Sometimes new N1QL query parameters are added before the SDK can be
updated or there are some undocumented parameters. To support this,
there should be a way to add raw parameters with a name and value to
a query.

MODIFICATIONS
-------------
- add RawParameter(name, value) to IQueryRequest and QueryRequest
- update QueryRequest store raw parameters and append to output sent to
  query service
- add unit test to ensure raw parameters are stored in output that is
  sent to query service

RESULT
------
Raw parameters with values can be added to n1ql queries.

NOTE: This has been added as an uncommitted feature until SDKs converge
on an agreed name (Raw vs Custom).

Change-Id: Ieeb9bb43c016ceba954c4324aca92975d5cb2580
Reviewed-on: http://review.couchbase.org/88922
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>