NCBC-1558: bootstrapping fails for memcd bucket when using couchbase protocol
Motivation
----------
If booting strapping to a memcached bucket and using the couchbase scheme
protocol, HttpClient will fail because it only supports HTTP and HTTPS
schemes. This patch replaces the couchbase protocol scheme with HTTP or
HTTPs.

Modifications
-------------
 - Added ReplaceCouchbaseSchemeWithHttp method to UriExtensions
 - Added code to replace couchbase schemes with HTTP schemes in
   HttpServerConfig
 - Added integration test for bootstrapping
 - Added unit tests to verify behaviour

Result
------
Bootstrapping to Memcached buckets using URI's with the couchbase scheme
will succeed.

Change-Id: I08e68d77cd103560825c70f5602da8af2856c3cb
Reviewed-on: http://review.couchbase.org/89455
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>