NCBC-1616: Track and log correlation details in Timeout Exceptions
MOTIVATION
----------
When a timeout occurs it would be beneficial to have additional details
related to the operation logged and added to the exception returned to
the application. The additional details should include: service type,
correlation ID, local endpoint, operation timeout and remote endpoint.

MODIFICATIONS
-------------
- Add CreateTimeoutException to ConnectionBase - this constructs the
  message that is both logged and used in the SendTimeoutException
- Update MultiplexingConnection (async & sync) to use
  CreateTimeoutException
- Update Connection & SslConnection sync paths to use
  CreateTimeoutException

RESULT
------
When a timeout happens, additional correlation details are added to the
SendTimeoutException to aid with cause analysis.

NOTE: Both Connection and SslConnection async paths do not currently
capture timeout exceptions.

Change-Id: I074c2a745f26c33030fbe64e30d5129ad937e2a0
Reviewed-on: http://review.couchbase.org/90060
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>