NCBC-1635: Default tracing / orphaned operation logging to disabled
MOTIVATION
----------
Tracing and orphaned response logging is a new feature for the SDK but
should not be enabled by default for 2.5 series of SDK.

MODIFICATIONS
-------------
- set both tracing and orphaned response to default to false in
  ClientConfiguration

RESULT
------
Both tracing and orphaned response logging is disabled for 2.5 series of
the SDK releases.

NOTE: From 2.6.0, tracing and orphaned response logging will be enabled
by default.

Change-Id: Ib6770a801bb6b3c1bc40c64d4d97db8270e51003
Reviewed-on: http://review.couchbase.org/90608
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>