NCBC-1639: Don’t add custom tags to NullSpans
MOTIVATION
----------
TracerExtensions add additional custom tags to spans during construction
of the spans. However, it is inefficient to add the custom tags if
tracing is disabled and the span type is NullSpan. If a NullSpan is
created, no additional tags should be added to the span and it should be
returned as fast as possible.

MODIFICATIONS
-------------
- Add checks for NullSpan when creating spans in TracerExtensions, if a
  NullSpan don't add extra tags and return span directly

RESULT
------
When a NullSpan is created using the ISpanBuilder, custom tags are now
added to the span.

Change-Id: Iea222de53c1a76baecc2154621fae4af4c7f786f
Reviewed-on: http://review.couchbase.org/90878
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>