NCBC-1643: Dependency on OpenTracing v0.10.0 throws exception in v2.5.6
Motivation
----------
The dependency on OpenTracing v0.10.0 throws a TypeInitializationException
with the message: A strongly-named assembly is required. This patch
configures the SDK to use a special OpenTracing nuget package that is
signed with the Couchbase SDK strong name keys.

Modifications
-------------
 - Add NuGet dependency to OpenTracing.Signed 0.10.2

Result
------
The SDK will pull and use a special signed version of OpenTracing.

Change-Id: I3c9ed9f3cf55e44d86b340065ff7bb259689e679
Reviewed-on: http://review.couchbase.org/91053
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>