NCBC-1591: Add Newtonsoft helper Unix epoch milliseconds attrs
Motivation
----------
Two very common ways of serializing date/times to JSON in Couchbase are
ISO 8601 and Unix epoch milliseconds.  The default behavior for the .NET
SDK is ISO 8601, but an easy and standardized way for developers to use
Unix timestamps would be useful.  This method could then be detected by
Linq2Couchbase to adjust query behavior.

Modifications
-------------
Added UnixMillisecondsConverter, a Newtonsoft JsonConverter, and
supporting unit tests.

Results
-------
Developers may apply the converter to any DateTime attribute, or
globally in their JsonSerializerSettings.

```cs
[JsonConverter(typeof(UnixMillisecondsConverter))]
public DateTime Timestamp { get; set; }
```

This results in both serializing and deserializing the property as
milliseconds since the Unix epoch.  Conversion to UTC is handled
automatically for local times.  Fractions of a millisecond are
preserved, up to the resolution of TimeSpan.TotalMilliseconds.

This attribute may then be detected by Linq2Couchbase to adjust query
generation for this properties.

Change-Id: I2b98799edbc0637b608cb55dc5cf6b53c33c1890
Reviewed-on: http://review.couchbase.org/91085
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>