NCBC-1645: Make SDK use OpenTracing.Signed v0.10.4
Motivation
----------
The assembly name for OpenTracing.Signed v0.10.2 does not match the
package name and will fail at runtime when .NET tries to load it. This
patch ensures that the SDK uses OpenTracing.Signed v0.10.4 which has the
correct assembly name.

Modification
------------
 - Change dependency on OpenTracing.Signed from v0.10.2 to v0.10.4

Result
------
The runtime will be able to find the OpenTracing.Signed assembly and load
it correctly.

Change-Id: I098972b4db0dda25cf5d9a2dc35af7c14c6f59ef
Reviewed-on: http://review.couchbase.org/91127
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Build Bot <build@couchbase.com>