NCBC-1651: SDK throws parse exception when geopoint field is read in response
Motivation
----------

If a geopoint type field is defined in an index, the response will fail
with: "Error reading string. Unexpected token: StartArray. Path
'fields.geo', line 1, position 413."

The reason is because the client expects the fields to be an array of
key/value strings; however, the value type for a geopoint Type is an array
of latitude/longitude which are numeric values:

"fields":{"geo":[9.15199995719641,41.38676998936262],"name":"Santateresa"}}

Modifications
-------------
 - Change SearchQueryRow.Fields from Dictionary<string, string> to
   Dictionary<string, dynamic>.

Results
-------
The field type will be dynammic and thus support all of different types
supported by field mapping: geopoint, string, boolean, and datetime. Note
that this is a breaking change, but I don't believe the other option of
converting everything to a string is the best solution.

Change-Id: I2947e21d5ab7166ee0e64de9dd8783cf4ef18a6f
Reviewed-on: http://review.couchbase.org/91381
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matthew D. Groves <me@mgroves.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>