NCBC-1666: Bad Authentication Causes Connection Leak
Motivation
----------
If the password for a user changes on the server, any clients that are
connected will not be able to create connections until the password is
also updated on the client. When this happens, the client will try to
reconnect every several seconds; if it cannot connect the socket must be
closed of else it will leak and the connection count on the server will
increase. This patch fixes a bug where the socket was not disposed when
authentication failed.

Modification
------------
Close of the connection if it cannot authenticate before throwing
exception.

Result
------
Server connections do not increase over time in the described situation.

Change-Id: I5bff93b0db1ce5b7ab1ba7653510ee2704fd6204
Reviewed-on: http://review.couchbase.org/92948
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
(cherry picked from commit 9e4746f3395c5330c6355ef0ddf8706011c19881)
Reviewed-on: http://review.couchbase.org/93225
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>