NCBC-1564: Update SearchException to use result message
MOTIVATION
----------
When creating a CouchbaseSearchResponseException, the
ISearchQueryResult's first error is used to generate the exception
message. However, this does not represent all the errors that have
occurred. The result.Message property should be used instead.

MODIFICATIONS
-------------
- use the result.Message property when building the error message

RESULT
------
The error message of the CouchbaseSearchResponseException is set using
the ISearchQuryResult.Message property not the first error.

Change-Id: I29ebbc32163e3b31053e18078c7928c38ce2a99a
Reviewed-on: http://review.couchbase.org/92978
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>