NCBC-1667: Check span latency is valid when processing
MOTIVATION
----------
When calculating the total duration of a span in the
ThresholdLoggingTracer, sometimes the span tag 'peer.latency' may have a
time precision suffix (us, ms, etc). These currently cause the
long.parse that tries to retrieve the value to break but should accept
the value and suffix and then calculate a real microsecond value.

MODIFICATIONS
-------------
- add TryConverToMicros utility function to TimeSpanExtensions
- add unit tests to verify behaviour (no suffix, us, ms and s)
- use TryConvertToMicros when processing spans

RESULT
------
When processing spans to calculate the total duration for a span
summary, durations with a suffix do not cause an error and are
calculated correctly.

Change-Id: I3f23f03b5afd408d91f06866767d857003fde0c1
Reviewed-on: http://review.couchbase.org/93068
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>