NCBC-1669: Consolidate ContinueOnAnyContext usage
MOTIVATION
----------
There is a utility function to configure a given task to not resume with
the current context, ContinueOnAnyContext. There are some usages in the
codebase that do not use this and it is desirable to have consistent
usage.

MODIFICATIONS
-------------
- update any direct usage of ConfigureAwait(false) to use the
  ConfigureOnAnyContext utility function

RESULT
------
The usage of ConfigureOnAnyContext is used consistently throughout the
codebase.

Change-Id: I4a585d84e919c0140af7562df4ac095278a9fbf5
Reviewed-on: http://review.couchbase.org/93193
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-on: http://review.couchbase.org/93249
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>