NCBC-1670: Dispose of syncstate pool when closing mux connection
MOTIVATION
----------
MultiplexingConnections contain a reusable queue of SyncStat's with
each SyncState having a AutoResetEvent which is not disposed correctly
when the connection is closed / disposed. The AutoResetEvent property
consumes system resources and should be disposed.

MODIFICATIONS
-------------
- add IDisposeable to IState
- add Dispose implementation to SyncState and AsyncState
- move states in flight clean up to outside of Socket != null check
- dequeue / dispose sync state when disposing from
  MultiplexingConnection

RESULT
------
The multiplexing connection's sync state pool is disposed correctly
when the connection is disposed.

Change-Id: I884c0ed45546568393c4a3c96ec3ad8f77611331
Reviewed-on: http://review.couchbase.org/93287
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>