NCBC-1674: Expose ITypeSerializer on IBucket
Motivation
----------
Eliminate the need for extensions such as Linq2Couchbase to instantiate
additional ITypeSerializer instances via ClientConfiguration when they
are accessing IBucket.

Modifications
-------------
Expose the ITypeSerializerProvider interface on CouchbaseBucket, and
return the ITypeSerizlizer instance from ITypeTranscoder.
implementations.

Results
-------
The shared instance of ITypeSerializer being used by the bucket is
available to external consumers of IBucket by type casting to
ITypeSerializerProvider.  Basic consumers of IBucket are not confused by
the new property because it is hidden.

Change-Id: I46eaf1da920e9202cceac235f83e276eda69040e
Reviewed-on: http://review.couchbase.org/93438
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>