NCBC-1679: Lock related errors not setting Exception and status
Motivation
----------
The KV Error Map feature that is available with Couchbase 5.X and a fixed
bug where the LOCKED status is returned when a key is locked opposed to
TEMP_FAIL require translation back to 4.X and lower semantics. This commit
reverts the client back to the pre-5.X error handling so backwards
compatibility is maintained with applications using older SDKs.

Modifications
-------------
 - Update OperationBase.GetResponseStatus to translate from LOCKED to
   TEMP_FAIL or KEY_EXISTS depending if operation is a mutation or not.
 - Update OperationHeaderExtensions.GetResponseStatus to allow the actual
   server status to be assigned as a result and then be translated by
   OperationBase.GetHeaderStatus.
 - Add newer response status types to ResponseStatus; some are not meant
   to be used directly by the Sdk but were added for consistency.
 - Add check of KV Error Map's error code to see if it indicates locked or
   not and if so add TemporaryLockFailureException to response.
 - Add unit tests

Result
------
Error handling for locked documents is similar to how it was for 4.X for
server 5.X+ versions.

See also NCBC-1500

Change-Id: I3f10cec23c52d87c31c8dbaa3190300479c6defc
Reviewed-on: http://review.couchbase.org/94091
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-on: http://review.couchbase.org/95137
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>