NCBC-1660: Handle FTS returning JSON for errors
Motivation
----------
From Couchbase Server 5.5, failed Search Query results will be a JSON
object with defined error message properties. Previously it was a raw
string message. The SDK should be able to handle both raw messages and
JSON objects and set the result error message accordingly.

Modifications
-------------
- add FailedSearchQueryResult class with Message and Status properties
- when processing an error, the SearchClient checks the content-type
  header to determine old or new behaviour
- add unit tests to verify new behaviour when receiving JSON content

Result
------
When processing a response for a failed search query, the content-type
is checked for appliation/json and if set, deserializes the content into
a known structure that is used to read the failure message. If the
content-type is not found, the whole content is set as the error
message.

Change-Id: I9fb6707c6d53cebc0de2bb74be7c447807ce4f44
Reviewed-on: http://review.couchbase.org/94298
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>