NCBC-1686: Update timeout messages to use JSON context
Motivation
----------
When a timeout occurs, operation specific contextual information is also
logged with the generic timeout message. Currently this information is
appended using a comma separated list, however, to improve readability
and to assist ingesting into other systems the information should be
presented as a compressed JSON object.

Modifications
-------------
- add OperationContext to hold contextual information (service type,
  correlation ID, bucket name, etc)
- create OperationContext when processing timeout for each service type
  and add whatever details are available
- add unit tests to verify OperationContext outputs expected JSON
  structure

Result
------
When a timeout occurs, the operation contextual information is appended
to the generic timeout message as a compressed JSON object.

Change-Id: I9a839f8128a7cf2be38bdf28bab6ef71edb93e1f
Reviewed-on: http://review.couchbase.org/94302
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>