NCBC-1699: Fix CouchbaseList to remove entries from backing document
MOTIVATION
----------
There is a bug where when an entry is tried to be removed from a
CouchbaseList collection, the item is not removed from the backing
Couchbase document. This commit ensures the entry is removed.

MODIFICATIONS
-------------
- add new internal RemoveAt function to remove the item in the backing
  document
- refactor existing methods to reuse GetItems().Contains(item)
- add unit tests to verify behaviour

RESULT
------
Items removed using the CouchbaseList collection are removed from the
Couchbase backing document in the cluster.

Change-Id: Ifa9891c42d25ea22aa0052eadd277be794b5615d
Reviewed-on: http://review.couchbase.org/95416
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>