NCBC-1717: Replace build badge with AppVeyor version
MOTIVATION
----------
The Jenkins build badge in the README can reflects all change set builds
and so can look like the project is failing when it's actually for an
un-merged commit. We should use the AppVeyor version instead which will
only include commits that have been merged into master.

MODIFICATIONS
-------------
- replace the Jenkins build badge in the README with the AppVeyor
  equivalent

RESULT
------
The build badge displayed in the README markdown now uses the AppVeyor
badge which reflects builds for merged commits only.

Change-Id: Ie5e7b8a33b6a40af6ceebf7b0ea09455f2749819
Reviewed-on: http://review.couchbase.org/96126
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>