NCBC-1736: Initialization fails when a node is unresponsive
MOTIVATION
----------
Whenever HttpServerConfig.Initialize encounters a problem with the connection with one of the servers, it should skip that server and try the next one. Because the DownloadString-method only throws HttpRequestExceptions, but the DownloadConfigs-method only handles BootstrapExceptions, this 'skip-and-try-the-next'-functionality does not work

MODIFICATIONS
-------------
De DownloadConfigs-method now also catches an exception of type HttpRequestException with the same handling as the catch-block for the BootstrapException. This was less invasive then rewriting the DownloadConfig-method to raise BootstrapExceptions instead of HttpRequestExceptions.
I think the catch-block for the BootstrapException-type can be removed, but I was not entirely sure.

RESULT
------
Whenever the initialization now encouters a non-responsive node, it will try to download the config from the next node in the list.

Change-Id: Ice6fc97404c0343f581bb1b4c63fe93a182d41a9
Reviewed-on: http://review.couchbase.org/96927
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>