NCBC-1725: Dispatch Increment via requst executor
MOTIVATION
----------
When dispatching an Increment operation during a rebalance, it is
normal for the server to return NotMyVBucket response codes. However,
these should not bubble out of the SDK and instead should be handled by
the SDK (retry, logging, set status codes, etc).

MODIFICATIONS
-------------
- dispatch synchronous Increment operations using the configured
  request executor instead of directly to a server in CouchbaseBucket

RESULT
------
Synchronous Increment operations are now dispatched using the internal
Request Executor instead of directly to a server. This ensures the same
failure handling processes are performed as with other operation types.

Change-Id: Ie10ba1ed70772bc11b63ddfab6a5f9cc57b016ea
Reviewed-on: http://review.couchbase.org/97401
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>