NCBC-1682: Use config.mgmtPort when streaming config
MOTIVATION
----------
When trying to get the streaming cluster configuration, the URI is first
checked if the couchbase:// protocol is used and switches it to either
http or https. However, the port is not automatically set so fails when
trying to connect. Instead, it should set the management port to the
config's management port.

MODIFICATIONS
-------------
- use the config.MgmtPort when replacing the couchbase:// protocol in
  UriExtensions
- add tests to verify behaviour in UriExtensionsTests

RESULT
------
When building a streaming URI that uses the couchbase:// protocol, the
ClientConfiguration mgmt port is used.

Change-Id: Iac5f005c2a2a430ac8a9899bb49cd7555cef20a3
Reviewed-on: http://review.couchbase.org/97555
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>