NCBC-1743: Fix collection modified error when building span data
MOTIVATION
----------
When a SpanSummary.Populate is called with an collection of spans, it
takes an IEnumerable<Span>. This should be copied to a local variable to
prevent an InvalidOperation if the collection is changed.

MODIFICATIONS
-------------
- ensure the span collection is fully enumerated before processing spans
  using ToList() in SpanSummary.PopulateSpan

RESULT
------
The collection is fully enumerated and will not be modified, therefore a
"System.InvalidOperationException: Collection was modified" cannot
occur.

Change-Id: I29d7826f9a86155f71c7bf8fc283f04ad725672b
Reviewed-on: http://review.couchbase.org/97791
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>