NCBC-1747: Don’t dispose cluster if not initialised during tests
MOTIVATION
----------
When executing ephemeral integration tests, the setup method checks if
the server support ephemeral buckets. If it does not, the tests are
skipped. However, the tear-down method is still called and does not
check if the cluster object is null before calling Dispose.

MODIFICATIONS
-------------
- check the cluster object is not null before calling Dispose

RESULT
------
The cluster object is checked to not be null before calling Dispose, and
therefore preventing a null reference exception.

Change-Id: I1e677615500ccc6ae70b27e189a4926e197c4d8d
Reviewed-on: http://review.couchbase.org/97999
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>