NCBC-1730: Client Failure at the beginning of restart server
Motivation
----------
When the server is shut down and there are pending operations, the socket
connection will timeout and a SendTimeoutExpiredException will be thrown
and then caught in exception handling. The status in this case was
incorrectly beng set to a generic ClientFailure which is incorrrect in
this case.

Modifications
-------------
Add specific error handling for SendTimeoutExpiredExceptions and setting
the status to OperationTimeout.

Result
------
If an operation timesout, the correct status of OperationTimeout will be
returned.

Change-Id: Id560c05626b543688073ba6acc0c072580316383
Reviewed-on: http://review.couchbase.org/98290
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>