NCBC-1759: Fix expiration bugs in CouchbaseBucket Increment and IncrementAsync
MOTIVATION
----------
Some of the Increment and IncrementAsync overloads set the timeout as
the expiration. Another overload set an incorrect timeout that would
cause TimeSpan.FromMinutes(5) to become 00:00:00.0000300. This would
then cause the expire time of the document to become zero (infinity)
instead of 5 minutes.

MODIFICATIONS
-------------
- Set GlobalTimeout for unit testing.
- Made some of the fields in Increment public readonly instead of a
  private field for unit testing.
- Refactored some of the code to make it more clear what is passed to
  the other overloads.
- Created separate files for testing the Increment and IncrementAsync
  methods.

RESULT
------
The Increment and IncrementAsync method will set the correct
expiration.

Change-Id: I474713a8f9a713d57658f0ff4c2d790ba3efbd23
Reviewed-on: http://review.couchbase.org/98792
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>