NCBC-1745: Add synchronisation locks around adding and getting sub-spans
MOTIVATION
----------
A Span has a list of sub-spans that us used to create a SpanSummary when
it has finished. However, when timeouts or latency is found in the
system, the parent span may finish and a summary be created before all
sub-spans are reported. This can cause the span list to be modified
during SpanSummary creation which uses a foreach and can throw an
exception.

MODIFICATIONS
-------------
- add a semaphoreslim to span that synchronises adding new spans and also
  returning the inner span list
- before adding a span, check if the end time stamp has been set and if
  so, don't add the span

RESULT
------
A Span's sub spans are now protected when adding or getting the list.

Change-Id: Ia50a799aa6baa5a1b08ce3eb4d6ec9454085e0c3
Reviewed-on: http://review.couchbase.org/98768
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>