NCBC-1757: use latest nodes/nodesext values when reusing node
Motivation
----------
When using an existing node and its resources during a cluster map change
ensure that the current node and nodeExt port values are updated in the
re-used node.

Modifications
-------------
 - Add Server.LoadNodeAdapter method which takes a INodeAdapter and
   refactor ctor to use it.
 - When a node is reused, call Server.LoadNodeAdapter and pass in the
   current cluster maps node and nodesExt values in MemcachedConfigContext
   and CouchbaseConfigContext
 - Add unit tests

Result
------
The current cluster maps ports will be used and the service will correctly
be located when mapping calls to a node.

Change-Id: I34773086879dc9163417710c8793d1d4ea60611a
Reviewed-on: http://review.couchbase.org/99108
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Build Bot <build@couchbase.com>