NCBC-1768: Add automatic Analytics request retries for certain error codes
MOTIVATION
----------
The Analytics team have indicated that certain combinations of query
status and error codes should automatically be retried and have been
added to the enhanced Analytics RFC.

MODIFICATIONS
-------------
- Update AnalyticsResult.ShouldRetry to return true for certain query
  status and error code combinations according to RFC.

RESULT
------
An analytics request is automatically retried if a certain error is
encountered. There is a maximum of 10 automatic retries, using an
exponential back-off.

Change-Id: I490f9177a5d23a2ba7389081f862e1fa9ab9f3a5
Reviewed-on: http://review.couchbase.org/99426
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>