NCBC-1762: Handle http status code 429 for FTS queries
MOTIVATION
----------
FTS queries may be returned with HTTP status code 429 if the search
nodes are over their query quote. In this scenario, the SDK should
automatically retry after a delay.

MODIFICATIONS
-------------
- Add internal HTTP status code to FTS result and set when
  processing result
- Update CouchbbaseRequestExecutor to check result.ShouldRetry
- Update SendWithRetryAsync for FTS to pass cancellation token to
  send delegate
- Add unit test to verify ShouldRetry works as expected

RESULT
------
When an FTS query is returned with HTTP status code 429 to indicate too
many requests, the SDK now retries with an exponential back-off with a
maximum of 10 retries.

Change-Id: I77baebc46e88ad589574763ebbaf0ebd52e25223
Reviewed-on: http://review.couchbase.org/99427
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>