NCBC-1694: Allow data structure ops to create doc if missing
MOTIVATION
----------
Data structure operations that add items to an underlying data structure
(map, list, set, queue), it should be able to create the document if
it's missing.

NOTE: We cannot rely on XATTR create_doc flag because it's only for CB
server 5.0+

MODIFICATIONS
-------------
- Create utility methods in CouchbaseBucket to execute subdoc operation
  for data structure and crate document if missing
- Add integration tests to verify behaviour

RESULT
------
Data structure calls that add items can now create the document if
it's missing.

Change-Id: Ie0c53ffadf878c224b7bbb758336a4ef43c6e5cb
Reviewed-on: http://review.couchbase.org/99544
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>