NCBC-1773: Add default tags to spans when not using built-in tracer
MOTIVATION
----------
When creating spans, default tags are added (eg span.client and
db.kind). However, because these tags are added within the SpanBuilder
implementation, which is internal to the Couchbase SDK for use with the
ThresholdLoggingTracer, these tags are not added when using an
alternative tracer (eg Jaeger). These spans should always be added,
irrespective of Tracer implementation.

MODIFICATIONS
-------------
- move the default tags initialised from within SpanBuilder to a new
  extension method in SpanBuilderExtensions
- use the new extension method within TracerExtensions when building
  spans
- add unit tests to verify spans are added as expected

RESULT
------
Spans that are created with an external Tracer implementation still
have all the default tags added during processing.

Change-Id: I3e6e5acd8b4026b756293f1820a09f5596d345b0
Reviewed-on: http://review.couchbase.org/99564
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>