NCBC-1774: Add document key span tag for KV operations
MOTIVATION
----------
When creating a span for a KV operation, it would be useful if the span
included the document key as a span tag. However, because document keys
can be considered sensitive information in some environments, it should
not be added if LogRedaction is enabled.

MODIFICATIONS
-------------
- Add DocumentKey member to CouchbaseTags
- Add DocumentKey span tag when creating KV spans, unless log redaction
  is enabled
- Update KV span tag tests to be check document key is present and be
  aware of log redaction level

RESULT
------
KV operations now have the document key added as a span tag unless log
redaction is enabled with the SDK.

Change-Id: I3f866bda6f4c561966dad0e62968bca48384afdd
Reviewed-on: http://review.couchbase.org/99570
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>