NCBC-1776: Fast-fail async KV operations based on operation timeout
MOTIVATION
----------
When a timeout is provided for async KV operations, the operation is
only checked for timeout after the SDK receives a response or if the
socket timeout (default of 15 seconds) is reached. It should fast-fail
client side and return a response to the application when the timeout is
reached and treat any server response as orphaned.

MODIFICATIONS
-------------
- register a delegate that sets the task completion source with a failed
  result when the cancellation token reaches it's timeout and removes
  the operation from the pending list for async KV opertions
- update CallbackFactory to use TryX methods to prevent a
  TaskAlreadyCompleted exception from being thrown if the TCS has
  already closed

RESULT
------
Async KV operations now fast-fail and return a failure result to the
appliation when a client-side timeout is provided.

Change-Id: I7860774e87f3134bd96bc00e86e6a9cf9764290c
Reviewed-on: http://review.couchbase.org/99795
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>