NCBC-1776: Async KV op timeout should use milliseconds not seconds
MOTIVATION
----------
Async KV operations use a timeout value with a CancellationToken,
however the value should be expressed in milliseconds, not seconds.
Also, the operations should use the provided timeout value instead of
the Default Operation Timeout.

MODIFICATIONS
-------------
- use the provided timeout for async KV operations instead of always
  using DefaultOperationTimeout
- replace TimeSpan extension method GetSeconds with GetMilliseconds and
  replace usage throughout codebase
- fix unit tests to assert timeout values are in milliseconds not
  seconds

RESULT
------
KV operations are expressed as milliseconds, not seconds and Async KV
operations use the provided timeout to fast-fail not the operation
default.

Change-Id: I5ed84bd40508106d11667d53471b3b9a6c679da8
Reviewed-on: http://review.couchbase.org/99796
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Build Bot <build@couchbase.com>